requirements.txt
python-dotenv==0.9.1
python-telegram-bot==12.8
requests==2.32.4
urllib3==1.26.20

